Data.Array.Accelerate.CUDA.Foreign

Description

Functions callable from foreign code

accelerateCreate :: Device -> ForeignContext -> IO AccHandle

Create an Accelerate handle given a device and a cuda context.

AccHandle accelerateCreate(int device, CUcontext ctx);

accelerateDestroy :: AccHandle -> IO ()

Releases all resources used by the accelerate library.

void accelerateDestroy(AccHandle hndl);

freeOutput :: Ptr OutputArray -> IO ()

Function callable from foreign code to free a OutputArray returned after executing an Accelerate computation.

Once freed, the device pointers associated with an array are no longer valid.

void freeOutput(OutputArray arr);

freeProgram :: StablePtr a -> IO ()

Free a compiled accelerate program.

void freeProgram(Program prg);

Exporting

exportAfun :: Name -> String -> Q [Dec]

Given the Name of an Accelerate function (a function of type ''Acc a -> Acc b'') generate a a function callable from foreign code with the second argument specifying it's name.

buildExported :: forall a b. (Arrays a, Arrays b) => AccHandle -> (Acc a -> Acc b) -> IO (StablePtr Afun)

Given a handle and an Accelerate function, generate an exportable version.

Types

type InputArray = (ShapeBuffer, DevicePtrBuffer)

The input required from foreign code.

type OutputArray = (ShapeBuffer, DevicePtrBuffer, StablePtr EArray)

A result array from an accelerate program.

type ShapeBuffer = Ptr CInt

A foreign buffer that represents a shape as an array of ints.

type DevicePtrBuffer = Ptr WordPtr

A buffer of device pointers

Backend representation

data CUDAForeignAcc f where

CUDA foreign Acc functions are just CIO functions.

Constructors

CUDAForeignAcc :: String -> (Stream -> as -> CIO bs) -> CUDAForeignAcc (as -> bs)

canExecuteAcc :: forall asm as bs. (Foreign asm, Typeable as, Typeable bs) => asm (as -> bs) -> Maybe (Stream -> as -> CIO bs)

Gives the executable form of a foreign function if it can be executed by the CUDA backend.

data CUDAForeignExp f where

CUDA foreign Exp functions consist of a list of C header files necessary to call the function and the name of the function to call.

Constructors

CUDAForeignExp :: IsScalar y => [String] -> String -> CUDAForeignExp (x -> y)

canExecuteExp :: forall asm x y. (Foreign asm, Typeable y, Typeable x) => asm (x -> y) -> Maybe ([String], String)

Gives the foreign function name as a string if it is a foreign Exp function for the CUDA backend.

Manipulating arrays

type family DevicePtrs e :: *

withDevicePtrs :: Array sh e -> Maybe Stream -> (DevicePtrs (EltRepr e) -> CIO b) -> CIO b

Get the raw CUDA device pointers associated with an array and call the given continuation.

indexArray :: (Shape dim, Elt e) => Array dim e -> Int -> CIO e

Read a single element from an array at the given row-major index. This is a synchronous operation.

useArray :: (Shape dim, Elt e) => Array dim e -> CIO ()

Upload an existing array to the device

useArrayAsync :: (Shape dim, Elt e) => Array dim e -> Maybe Stream -> CIO ()

peekArray :: (Shape dim, Elt e) => Array dim e -> CIO ()

peekArrayAsync :: (Shape dim, Elt e) => Array dim e -> Maybe Stream -> CIO ()

pokeArray :: (Shape dim, Elt e) => Array dim e -> CIO ()

pokeArrayAsync :: (Shape dim, Elt e) => Array dim e -> Maybe Stream -> CIO ()

copyArray :: (Shape dim, Elt e) => Array dim e -> Array dim e -> CIO ()

Copy data between two device arrays. The operation is asynchronous with respect to the host, but will never overlap kernel execution.

copyArrayAsync :: (Shape dim, Elt e) => Array dim e -> Array dim e -> Maybe Stream -> CIO ()

allocateArray :: (Shape dim, Elt e) => dim -> CIO (Array dim e)

fromFunction :: (Shape sh, Elt e) => sh -> (sh -> e) -> CIO (Array sh e)

Running IO actions in an Accelerate context

data CIO a

data Stream :: *

A processing stream. All operations in a stream are synchronous and executed in sequence, but operations in different non-default streams may happen out-of-order or concurrently with one another.

Use Events to synchronise operations between streams.

liftIO :: MonadIO m => forall a. IO a -> m a

Lift a computation from the IO monad.

inContext :: Context -> IO a -> IO a

Run an IO action within the given Acclerate context

inDefaultContext :: IO a -> IO a

Run an IO action in the default Acclerate context